requirements.txt
imutils==0.5.4
matplotlib==3.6.2
numpy==1.23.4
opencv_python==4.9.0.80
Pillow==9.3.0
Pillow==10.3.0
skimage==0.0
streamlit==1.34.0
